sdk: refactor client options into builder pattern
Ref https://github.com/rust-nostr/nostr/issues/920

Pull-Request: https://github.com/rust-nostr/nostr/pull/1241
